delete the useless post folder and move hashtag post for the better visibility
