+++ src/api-post/prepareRoutes.js
@@ -1,8 +1,8 @@
-import prepareRoutesForYou from "./src/routes/posts/ForYouRoute.js"
-import prepareRoutesSubscribed from "./src/routes/posts/SubscribedRoute.js"
-import prepareRoutesLike from "./src/routes/posts/LikeRoute.js"
-import prepareRoutesReport from "./src/routes/posts/ReportRoute.js"
-import prepareRoutesDeletePost from "./src/routes/posts/DeletePostRoute.js"
+import prepareRoutesForYou from "./src/routes/ForYouRoute.js"
+import prepareRoutesSubscribed from "./src/routes/SubscribedRoute.js"
+import prepareRoutesLike from "./src/routes/LikeRoute.js"
+import prepareRoutesReport from "./src/routes/ReportRoute.js"
+import prepareRoutesDeletePost from "./src/routes/DeletePostRoute.js"
 import prepareRoutesCreatePost from "./src/routes/CreatePostRoute.js"
 
 const prepareRoutes = ctx => {
